move all low level functions to lib.rs
