--- file_store/src/file_store.rs
@@ -1,178 +1,0 @@
-use crate::{BytesMutStream, Error, FileInfo, FileInfoStream, Result};
-use aws_sdk_s3::{primitives::ByteStream, Client};
-use aws_smithy_types_convert::stream::PaginationStreamExt;
-use chrono::{DateTime, Utc};
-use futures::{future, stream, FutureExt, StreamExt, TryFutureExt, TryStreamExt};
-use std::path::Path;
-
-pub struct FileData {
-    pub info: FileInfo,
-    pub stream: BytesMutStream,
-}
-
-pub fn list_files<A, B>(
-    client: &Client,
-    bucket: &str,
-    prefix: &str,
-    after: A,
-    before: B,
-) -> FileInfoStream
-where
-    A: Into<Option<DateTime<Utc>>> + Copy,
-    B: Into<Option<DateTime<Utc>>> + Copy,
-{
-    let file_type = prefix.to_string();
-    let before = before.into();
-    let after = after.into();
-
-    client
-        .list_objects_v2()
-        .bucket(bucket)
-        .prefix(file_type.to_string())
-        .set_start_after(after.map(|dt| FileInfo::from((file_type, dt)).into()))
-        .into_paginator()
-        .send()
-        .into_stream_03x()
-        .map_ok(|page| stream::iter(page.contents.unwrap_or_default()).map(Ok))
-        .map_err(|err| Error::from(aws_sdk_s3::Error::from(err)))
-        .try_flatten()
-        .try_filter_map(|file| future::ready(FileInfo::try_from(&file).map(Some)))
-        .try_filter(move |info| future::ready(after.is_none_or(|v| info.timestamp > v)))
-        .try_filter(move |info| future::ready(before.is_none_or(|v| info.timestamp <= v)))
-        .boxed()
-}
-
-pub async fn list_all_files<A, B>(
-    client: &Client,
-    bucket: &str,
-    file_type: &str,
-    after: A,
-    before: B,
-) -> Result<Vec<FileInfo>>
-where
-    A: Into<Option<DateTime<Utc>>> + Copy,
-    B: Into<Option<DateTime<Utc>>> + Copy,
-{
-    list_files(client, bucket, file_type, after, before)
-        .try_collect()
-        .await
-}
-
-pub async fn put_file(client: &Client, bucket: &str, file: &Path) -> Result {
-    let byte_stream = ByteStream::from_path(&file)
-        .await
-        .map_err(|_| Error::not_found(format!("could not open {}", file.display())))?;
-    poc_metrics::record_duration!(
-        "file_store_put_duration",
-        client
-            .put_object()
-            .bucket(bucket)
-            .key(file.file_name().map(|name| name.to_string_lossy()).unwrap())
-            .body(byte_stream)
-            .content_type("application/octet-stream")
-            .send()
-            .map_ok(|_| ())
-            .map_err(Error::s3_error)
-            .await
-    )
-}
-
-pub async fn remove_file(client: &Client, bucket: &str, key: &str) -> Result {
-    poc_metrics::record_duration!(
-        "file_store_remove_duration",
-        client
-            .delete_object()
-            .bucket(bucket)
-            .key(key)
-            .send()
-            .map_ok(|_| ())
-            .map_err(Error::s3_error)
-            .await
-    )
-}
-
-pub async fn get_raw_file<K>(client: &Client, bucket: &str, key: K) -> Result<ByteStream>
-where
-    K: Into<String>,
-{
-    get_byte_stream(client.clone(), bucket.to_string(), key).await
-}
-
-pub async fn get_file<K>(client: &Client, bucket: &str, key: K) -> Result<BytesMutStream>
-where
-    K: Into<String>,
-{
-    Ok(stream_source(get_raw_file(client, bucket, key).await?))
-}
-
-pub fn source_files(client: &Client, bucket: &str, infos: FileInfoStream) -> BytesMutStream {
-    let bucket = bucket.to_string();
-    let client = client.clone();
-    infos
-        .map_ok(move |info| get_byte_stream(client.clone(), bucket.clone(), info.key))
-        .try_buffered(2)
-        .flat_map(|stream| match stream {
-            Ok(stream) => stream_source(stream),
-            Err(err) => stream::once(async move { Err(err) }).boxed(),
-        })
-        .fuse()
-        .boxed()
-}
-
-pub fn source_files_unordered(
-    client: &Client,
-    bucket: &str,
-    workers: usize,
-    infos: FileInfoStream,
-) -> BytesMutStream {
-    let bucket = bucket.to_string();
-    let client = client.clone();
-    infos
-        .map_ok(move |info| get_byte_stream(client.clone(), bucket.clone(), info.key))
-        .try_buffer_unordered(workers)
-        .flat_map(|stream| match stream {
-            Ok(stream) => stream_source(stream),
-            Err(err) => stream::once(async move { Err(err) }).boxed(),
-        })
-        // .fuse()
-        .boxed()
-}
-
-pub async fn stream_single_file(
-    client: &Client,
-    bucket: &str,
-    file_info: FileInfo,
-) -> Result<BytesMutStream> {
-    get_byte_stream(client.clone(), bucket.to_string(), file_info)
-        .await
-        .map(stream_source)
-}
-
-pub fn stream_source(stream: ByteStream) -> BytesMutStream {
-    use async_compression::tokio::bufread::GzipDecoder;
-    use tokio::io::BufReader;
-    use tokio_util::codec::{length_delimited::LengthDelimitedCodec, FramedRead};
-
-    Box::pin(
-        FramedRead::new(
-            GzipDecoder::new(BufReader::new(stream.into_async_read())),
-            LengthDelimitedCodec::new(),
-        )
-        .map_err(Error::from),
-    )
-}
-
-async fn get_byte_stream<K>(client: Client, bucket: String, key: K) -> Result<ByteStream>
-where
-    K: Into<String>,
-{
-    client
-        .get_object()
-        .bucket(bucket)
-        .key(key)
-        .send()
-        .map_ok(|output| output.body)
-        .map_err(Error::s3_error)
-        .fuse()
-        .await
-}
